+++ utils/constants.py
@@ -1,5 +1,5 @@
 from pathlib import Path
-APP_VERSION = "0.2.140"
+APP_VERSION = "0.2.141"
 
 
 # Player tags